+++ app/app.py
@@ -1,10 +1,13 @@
 import falcon
-from resources.service import Predict, Health
+from resources.model import Predict, Health
 
-from middleware.logger import logger
+import logging, sys
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG,
+                    format=f"%(asctime)s - [%(levelname)s] - %(name)s - (%(filename)s).%(funcName)s(%(lineno)d) - %(message)s")
+logger = logging.getLogger(__name__)
 
 api = falcon.App()
 
-api.add_route("/action2", Predict())
+api.add_route('/predict/{task_id}', Predict())
 api.add_route("/health", Health())
 logger.debug('service started')
